Dededuplicate 2 functions in tools: _iterate _get_groups
resolves #350
