+++ project_generator/tools/eclipse.py
@@ -71,42 +71,6 @@ class EclipseGnuARM(Tool, Exporter, Builder):
                 new_data['groups'][group].append(new_file)
         new_data['groups'][group] = sorted(new_data['groups'][group], key=lambda x: x['name'].lower())
 
-#TODO: eliminate this duplicate in many tools. Same applies to _iterate
-    def _get_groups(self, data):
-        """ Get all groups defined. """
-        groups = []
-        for attribute in SOURCE_KEYS:
-            for k, v in data[attribute].items():
-                if k == None:
-                    k = 'Sources'
-                if k not in groups:
-                    groups.append(k)
-            for k, v in data['include_files'].items():
-                if k == None:
-                    k = 'Includes'
-                if k not in groups:
-                    groups.append(k)
-        return groups
-
-    def _iterate(self, data, expanded_data):
-        """ Iterate through all data, store the result expansion in extended dictionary. """
-        for attribute in SOURCE_KEYS:
-            for k, v in data[attribute].items():
-                if k == None:
-                    group = 'Sources'
-                else:
-                    group = k
-                self._expand_data(data[attribute], expanded_data, attribute, group)
-        for k, v in data['include_files'].items():
-            if k == None:
-                group = 'Includes'
-            else:
-                group = k
-            self._expand_data(data['include_files'], expanded_data, attribute, group)
-
-        # sort groups
-        expanded_data['groups'] = OrderedDict(sorted(expanded_data['groups'].items(), key=lambda t: t[0]))
-
     def export_workspace(self):
         logger.debug("Current version of CoIDE does not support workspaces")
