+++ project_generator/tools/tool.py
@@ -14,10 +14,13 @@
 
 import os
 import logging
+from collections import OrderedDict
 
 from os.path import join, dirname, abspath
 from jinja2 import Template, FileSystemLoader
 from jinja2.environment import Environment
+
+from ..util import SOURCE_KEYS
 
 logger = logging.getLogger('progen.tools')
 
@@ -103,3 +106,38 @@ class Exporter(object):
     @staticmethod
     def is_supported_by_default(target):
         return False
+
+    def _get_groups(self, data):
+        """ Get all groups defined """
+        groups = []
+        for attribute in SOURCE_KEYS:
+            for k, v in data[attribute].items():
+                if k == None:
+                    k = 'Sources'
+                if k not in groups:
+                    groups.append(k)
+            for k, v in data['include_files'].items():
+                if k == None:
+                    k = 'Includes'
+                if k not in groups:
+                    groups.append(k)
+        return groups
+
+    def _iterate(self, data, expanded_data):
+        """ _Iterate through all data, store the result expansion in extended dictionary """
+        for attribute in SOURCE_KEYS:
+            for k, v in data[attribute].items():
+                if k == None:
+                    group = 'Sources'
+                else:
+                    group = k
+                self._expand_data(data[attribute], expanded_data, attribute, group)
+        for k, v in data['include_files'].items():
+            if k == None:
+                group = 'Includes'
+            else:
+                group = k
+            self._expand_data(data['include_files'], expanded_data, attribute, group)
+
+        # sort groups
+        expanded_data['groups'] = OrderedDict(sorted(expanded_data['groups'].items(), key=lambda t: t[0]))
